+++ lib/compress_result.dart
@@ -1,7 +1,0 @@
-part of 'flutter_video_compress.dart';
-
-class CompressResult {
-  final String path;
-  final bool isCancel;
-  const CompressResult({@required this.path, @required this.isCancel});
-}