+++ lib/model/media_info.dart
@@ -1,0 +1,55 @@
+part of flutter_video_compress;
+
+class MediaInfo {
+  String path;
+  String title;
+  String author;
+  int width;
+  int height;
+
+  /// bytes
+  int filesize; // filesize
+  /// microsecond
+  double duration;
+  bool isCancel;
+  File file;
+
+  MediaInfo({
+    @required this.path,
+    this.title,
+    this.author,
+    this.width,
+    this.height,
+    this.filesize,
+    this.duration,
+    this.isCancel,
+    this.file,
+  });
+
+  MediaInfo.fromJson(Map<String, dynamic> json) {
+    path = json['path'];
+    title = json['title'];
+    author = json['author'];
+    width = json['width'];
+    height = json['height'];
+    filesize = json['filesize'];
+    duration = double.tryParse('${json['duration']}');
+    isCancel = json['isCancel'];
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = Map<String, dynamic>();
+    data['path'] = this.path;
+    data['title'] = this.title;
+    data['author'] = this.author;
+    data['width'] = this.width;
+    data['height'] = this.height;
+    data['filesize'] = this.filesize;
+    data['duration'] = this.duration;
+    if (this.isCancel != null) {
+      data['isCancel'] = this.isCancel;
+    }
+    data['file'] = File(this.path);
+    return data;
+  }
+}