+++ lib/src/video_quality.dart
@@ -1,0 +1,3 @@
+part of flutter_video_compress;
+
+enum VideoQuality { LowQuality, MediumQuality, HighestQuality }